Update requirements management to use pip-compile (#2)
This switches to using pip-tools to do requirements management. Without
this we get kind of a twisty mess of installed requirements and led to
fun things like having mixed versions of pydantic. This pins all of our
requirements and also expresses that we don't want pydantic 2.x yet.
Also added a gradle task to build the requirements.txt file since we
want to exclude "py" due to a bogus CVE.

We also switch to using pyproject.toml to express the project
configuration for setuptools (since this is the new, new standard).

No issue logged...