Add a PDF that we can use in our testing of the AI Assistant (#18)
We are doing this here because it is a public repo. As a result we get a stable, https based URI for the PDF which we need for testing. We also get a redirect if we use the "http" form of the URI, which is also something we test for.